Replace to Player class
